introduced abstract provider test
